Main UI (#7)
* feat: 기초 틀 생성

* refactor: ui 패키지 client로 이동 및 README.md 업데이트

* refactor: ChatUI 사이드바 제거 및 크기 조정

* feat: 사이드바 버튼을 이용한 화면 전환 기능 추가
