+++ README.md
@@ -7,14 +7,11 @@ Java를 이용한 암호화 메신저 서비스입니다.
 
 ## ✨ Features
 
-> **아직 확정되지 않았습니다**
 - 1:1 채팅(그룹 채팅)
-  - 채팅 이름, 아바타 설정
   - 텍스트 메세지
     - 메세지 입력창 UI, UX
     - 보낸 시간 표시
     - 채팅 내역은 DB에 **암호화**하여 저장
-  - 유저 온/오프라인 표시
   - 읽지 않은 유저 수 카운트
 - 회원가입
   - 아이디 중복 확인
@@ -24,7 +21,6 @@ Java를 이용한 암호화 메신저 서비스입니다.
 - 로그인/로그아웃
   - 자동 로그인
   - 비밀번호 찾기
-
 
 ## ⚙️ Build & Run
 ```bash
@@ -51,29 +47,14 @@ secure-messenger/
 ├── src/
 │   ├── main/
 │   │   ├── java/
-│   │   │   ├── client/               # 클라이언트 애플리케이션
-│   │   │   │   └── ClientMain.java
+│   │   │   ├── client/               # 클라이언트 애플리케이션 (UI 포함)
 │   │   │   ├── server/               # 서버 애플리케이션
-│   │   │   │   ├── ServerMain.java
-│   │   │   │   └── ClientSession.java
-│   │   │   ├── crypto/               # 암호화 로직 (RSA 키 생성, 암복호화)
-│   │   │   │   ├── KeyGenerator.java
-│   │   │   │   ├── Encryptor.java
-│   │   │   │   └── Decryptor.java
-│   │   │   ├── common/               # 공통 데이터 모델 및 유틸리티
-│   │   │   │   ├── Message.java
-│   │   │   │   ├── Protocol.java
-│   │   │   │   └── Logger.java
-│   │   │   ├── ui/                   # ui
-│   │   │   │   ├── ChatUI.java
-│   │   │   │   ├── LoginUI.java
-│   │   │   │   └── SignUpUI.java
+│   │   │   ├── crypto/               # 암호화 로직
+│   │   │   ├── common/               # 공통 클래스
 │   └── main/
-│       └── resources/
-│           ├── schema.sql            # MySQL 테이블 생성 스크립트
-│           └── config.properties     # DB 설정 등 설정파일
-├── build.gradle                      # Gradle 빌드 설정 파일
-├── settings.gradle                   # Gradle 프로젝트 설정
+│       └── resources/                # 리소스
+├── build.gradle                      
+├── settings.gradle                   
 └── .gitignore
 
 ```
