+++ src/main/java/client/ClientMain.java
@@ -1,9 +1,11 @@
 package client;
 
-import ui.ChatUI;
+import client.ui.ChatUI;
+import client.ui.MainUI;
 
 public class ClientMain {
     public static void main(String[] args) {
-        ChatUI chatUI = new ChatUI("이름이너무길다");
+//        ChatUI chatUI = new ChatUI(null);
+        MainUI mainUI = new MainUI(null);
     }
 }
