+++ src/main/java/client/ui/LoginUI.java
@@ -1,0 +1,13 @@
+package client.ui;
+
+import javax.swing.*;
+
+public class LoginUI extends JFrame {
+    // TODO 1. 로그인 UI 만들기
+    // TODO 2. 아이디 찾기, 비밀번호 찾기 UI 만들기
+    /**
+     * LoginUI 생성
+     */
+    public LoginUI() {
+    }
+}
