+++ src/main/java/client/ui/panel/SettingsPanel.java
@@ -1,0 +1,12 @@
+package client.ui.panel;
+
+import javax.swing.*;
+
+public class SettingsPanel extends JPanel {
+    public SettingsPanel() {
+        JLabel label = new JLabel("Settings");
+        label.setHorizontalAlignment(SwingConstants.CENTER);
+        label.setVerticalAlignment(SwingConstants.CENTER);
+        add(label);
+    }
+}
